rom: fix clippy warnings and remove panic paths (#1484)
Fix all per-variant clippy lints and eliminate panic symbols from
the FPGA flash driver. Enable per-variant clippy and panic-check
for all 11 ROM variants in precheckin.

Add `cargo xtask panic-check` subcommand for ad-hoc use.

+++ platforms/emulator/rom/src/riscv.rs
@@ -49,7 +49,6 @@ use caliptra_mcu_romtime::HexWord;
 unsafe fn init_static<T>(s: *mut core::mem::MaybeUninit<T>, value: T) -> &'static mut T {
     (*s).write(value)
 }
-
 use zerocopy::{transmute, FromBytes, IntoBytes};
 
 /// DOT recovery handler using MCI mbox0.
@@ -768,7 +767,7 @@ pub extern "C" fn rom_entry() -> ! {
 
     caliptra_mcu_romtime::println!(
         "[mcu-rom] Jumping to firmware at {}",
-        HexWord(MCU_MEMORY_MAP.sram_offset as u32)
+        HexWord(MCU_MEMORY_MAP.sram_offset)
     );
     exit_rom();
 }
